x86/mm/kmmio: Fix mmiotrace for page unaligned addresses
If something calls ioremap(0x.., 0x200) it may end up returning 0x..200.
This lead to a probe registered on exactly the returned address, but the
entire page was armed for mmiotracing.

On iounmap(0x..) the address passed to unregister_kmmio_probe was already
page aligned leading to a complete freeze of the machine.

We should always page align addresses while (un)registerung mappings,
because the mmiotracer works on top of pages, not mappings. We still keep
track of the probes based on their real addresses and lengths though,
because that's the significant part while handling the page faults.

Also move the call to mmiotrace_iounmap prior page aligning the address,
so that all probes are unregistered properly, otherwise the kernel ends up
failing memory allocations randomly after disabling the mmiotracer.

Cc: Steven Rostedt <rostedt@goodmis.org>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Pekka Paalanen <ppaalanen@gmail.com>
Cc: linux-kernel@vger.kernel.org
Cc: nouveau@lists.freedesktop.org
Cc: x86@kernel.org
Acked-by: Pekka Paalanen <ppaalanen@gmail.com>
Signed-off-by: Karol Herbst <kherbst@redhat.com>