WIPdrm/nouveau/kms/nv50-: skip bios parsing entirely
We shouldn't need this now on >=NV50, and Ada will fail unless it's
removed as the DCB table appears to be moved/missing.

This is hackish.  I want to rework and cleanup this whole series a bit,
and will handle things better then.

WIP: DAC testval -> nvkm
WIP: LVDS stuff still needs moving to nvkm

Signed-off-by: Ben Skeggs <bskeggs@redhat.com>